Parented player to a prefab gameObject
