+++ com.unity.cinemachine/Samples~/Input System/Split Screen Multiplayer/PlayerWithRig.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4d8d7a9a98d3f4ac2967d48094ea010f
+guid: 33f863ca3153d440d84eccc416174573
 PrefabImporter:
   externalObjects: {}
   userData:
